Provision RStudio environment to use for 1) final preparations for the 2019 miniDREAM course 2) production environment for the miniDREAM course
Instance will run RStudio along with its public web interface (hence we don't
want to run it on scicomp...) allowing access to an R environment
for about 20 users on the same machine.

Users will perform calculations on datasets, loading potentially dozens
of GBs of data in memory, concurrently.

We would like to keep the previous sandbox test instance running at 10.23.12.137
active as well, since we might want to experiment with it.